@@ -1,4 +1,4 @@
-class CreateAdminService
+class CreateAdminService < ApplicationService
   def call
     user = User.find_or_create_by!(email: Rails.application.secrets.admin_email) do |user|
         user.password = Rails.application.secrets.admin_password
