@@ -8,9 +8,10 @@ Bundler.require(*Rails.groups)
 
 module Detail
   class Application < Rails::Application
+    config.load_defaults 7.1
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
-    config.active_job.queue_adapter = :delayed_job
+    config.active_job.queue_adapter = :sidekiq
   end
 end
