@@ -1,94 +1,67 @@
-# config valid for current version and patch releases of Capistrano
-lock "~> 3.14.0"
+# lock "~> 3.14.0"
 
-# set :application, "my_app_name"
-# set :repo_url, "git@example.com:me/my_repo.git"
+# set :application, 'detail'
+# set :repo_url, 'git@github.com:stap780/detail.git'
+# set :deploy_to, '/var/www/detail'
+# set :linked_files, fetch(:linked_files, []).push('config/database.yml', 'config/secrets.yml')
+# set :linked_dirs, fetch(:linked_dirs, []).push('log', 'tmp/pids', 'tmp/cache', 'tmp/sockets', 'vendor/bundle', 'public')
+# set :format, :pretty
+# set :log_level, :info
+# set :whenever_identifier, ->{ "#{fetch(:application)}_#{fetch(:stage)}" }
+# set :delayed_job_workers, 1
+# set :delayed_job_roles, [:app]
+# set :delayed_job_pid_dir, '/tmp'
 
-# Default branch is :master
-# ask :branch, `git rev-parse --abbrev-ref HEAD`.chomp
+lock "~> 3.18.1"
 
-# Default deploy_to directory is /var/www/my_app_name
-# set :deploy_to, "/var/www/my_app_name"
+server '104.131.21.204', user: 'deploy', roles: %w{app db web}
 
-# Default value for :format is :airbrussh.
-# set :format, :airbrussh
-
-# You can configure the Airbrussh format using :format_options.
-# These are the defaults.
-# set :format_options, command_output: true, log_file: "log/capistrano.log", color: :auto, truncate: :auto
-
-# Default value for :pty is false
-# set :pty, true
-
-# Default value for :linked_files is []
-# append :linked_files, "config/database.yml"
-
-# Default value for linked_dirs is []
-# append :linked_dirs, "log", "tmp/pids", "tmp/cache", "tmp/sockets", "public/system"
-
-# Default value for default_env is {}
-# set :default_env, { path: "/opt/ruby/bin:$PATH" }
-
-# Default value for local_user is ENV['USER']
-# set :local_user, -> { `git config user.name`.chomp }
-
-# Default value for keep_releases is 5
-# set :keep_releases, 5
-
-# Uncomment the following to require manually verifying the host key before first deploy.
-# set :ssh_options, verify_host_key: :secure
-set :application, 'detail'
-set :repo_url, 'git@github.com:stap780/detail.git'
-set :deploy_to, '/var/www/detail'
-set :linked_files, fetch(:linked_files, []).push('config/database.yml', 'config/secrets.yml')
-set :linked_dirs, fetch(:linked_dirs, []).push('log', 'tmp/pids', 'tmp/cache', 'tmp/sockets', 'vendor/bundle', 'public')
-set :format, :pretty
-set :log_level, :info
-set :whenever_identifier, ->{ "#{fetch(:application)}_#{fetch(:stage)}" }
-set :delayed_job_workers, 1
-set :delayed_job_roles, [:app]
-set :delayed_job_pid_dir, '/tmp'
+set :application, "detail"
+set :repo_url, "git@github.com:stap780/#{fetch(:application)}.git"
 
 
+set :user, 'deploy'
 
+set :branch, "master"
+set :pty,             true
+set :stage,           :production
+set :deploy_to,       "/var/www/#{fetch(:application)}"
+set :puma_access_log, "#{release_path}/log/puma.access.log"
+set :puma_error_log,  "#{release_path}/log/puma.error.log"
+set :ssh_options,     { forward_agent: true, user: fetch(:user), keys: %w(~/.ssh/id_rsa.pub) }
+# set :puma_enable_socket_service, true
 
-# set :application, 'my_app_name'
-# set :repo_url, 'git@example.com:me/my_repo.git'
-#
-# # ask :branch, proc { `git rev-parse --abbrev-ref HEAD`.chomp }
-#
-# # set :deploy_to, '/var/www/my_app'
-# # set :scm, :git
-#
-# # set :format, :pretty
-# # set :log_level, :debug
-# # set :pty, true
-#
-# # set :linked_files, %w{config/database.yml}
-# # set :linked_dirs, %w{bin log tmp/pids tmp/cache tmp/sockets vendor/bundle public/system}
-#
-# # set :default_env, { path: "/opt/ruby/bin:$PATH" }
-# # set :keep_releases, 5
-#
-# namespace :deploy do
-#
-#   desc 'Restart application'
-#   task :restart do
-#     on roles(:app), in: :sequence, wait: 5 do
-#       # Your restart mechanism here, for example:
-#       # execute :touch, release_path.join('tmp/restart.txt')
-#     end
-#   end
-#
-#   after :restart, :clear_cache do
-#     on roles(:web), in: :groups, limit: 3, wait: 10 do
-#       # Here we can do anything such as:
-#       # within release_path do
-#       #   execute :rake, 'cache:clear'
-#       # end
-#     end
-#   end
-#
-#   after :finishing, 'deploy:cleanup'
-#
-# end
+append :linked_files, "config/master.key", "config/database.yml", "config/sidekiq.yml"
+append :linked_dirs, "log", "tmp/pids", "tmp/cache", "public", 'tmp/sockets', 'vendor/bundle', 'lib/tasks', 'lib/drop', 'storage'
+
+namespace :puma do
+    desc 'Create Directories for Puma Pids and Socket'
+    task :make_dirs do
+      on roles(:app) do
+        execute "mkdir #{shared_path}/tmp/sockets -p"
+        execute "mkdir #{shared_path}/tmp/pids -p"
+      end
+    end
+  
+    before 'deploy:starting', 'puma:make_dirs'
+end
+
+namespace :deploy do
+  desc "Make sure local git is in sync with remote."
+  task :check_revision do
+    on roles(:app) do
+      # Update this to your branch name: master, main, etc. Here it's master
+      unless `git rev-parse HEAD` == `git rev-parse origin/main`
+        puts "WARNING: HEAD is not the same as origin/main"
+        puts "Run `git push` to sync changes."
+        exit
+      end
+    end
+  end
+
+  before :starting,     :check_revision
+  after  :finishing,    :compile_assets
+  after  :finishing,    :cleanup
+  
+end
+  
