@@ -23,8 +23,7 @@ Rails.application.configure do
   else
     config.action_controller.perform_caching = false
 
-    # config.cache_store = :null_store
-    config.cache_store = :memory_store
+    config.cache_store = :null_store
   end
 
   # Don't care if the mailer can't send.
@@ -41,7 +40,7 @@ Rails.application.configure do
   # Debug mode disables concatenation and preprocessing of assets.
   # This option may cause significant delays in view rendering with a large
   # number of complex assets.
-  config.assets.debug = true
+  # config.assets.debug = true
 
   config.action_mailer.smtp_settings = {
     tls: true,
@@ -69,5 +68,5 @@ Rails.application.configure do
 
   # Use an evented file watcher to asynchronously detect changes in source code,
   # routes, locales, etc. This feature depends on the listen gem.
-  config.file_watcher = ActiveSupport::EventedFileUpdateChecker
+  # config.file_watcher = ActiveSupport::EventedFileUpdateChecker
 end
