@@ -39,10 +39,16 @@ every 1.day, :at => '16:00' do
   runner "Product.import"
 end
 
+# every 1.day, :at => '23:00' do
+#   runner "Kare.import"
+# end
+
 every 1.day, :at => '23:00' do
-  runner "Kare.import"
+  runner "Kare.pars"
 end
-every 1.day, :at => '23:10' do
+
+
+every 1.day, :at => '03:10' do
   runner "Kare.csv_param"
 end
 
