@@ -2,79 +2,81 @@
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# Note that this schema.rb definition is the authoritative source for your
-# database schema. If you need to create the application database on another
-# system, you should be using db:schema:load, not running all the migrations
-# from scratch. The latter is a flawed and unsustainable approach (the more migrations
-# you'll amass, the slower it'll run and the greater likelihood for issues).
+# This file is the source Rails uses to define your schema when running `bin/rails
+# db:schema:load`. When creating a new database, `bin/rails db:schema:load` tends to
+# be faster and is potentially less error prone than running all of your
+# migrations from scratch. Old migrations may fail to apply correctly if those
+# migrations use external dependencies or application code.
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20210309132849) do
-
+ActiveRecord::Schema[7.1].define(version: 2024_06_23_151335) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
-  create_table "delayed_jobs", force: :cascade do |t|
-    t.integer  "priority",   default: 0, null: false
-    t.integer  "attempts",   default: 0, null: false
-    t.text     "handler",                null: false
-    t.text     "last_error"
-    t.datetime "run_at"
-    t.datetime "locked_at"
-    t.datetime "failed_at"
-    t.string   "locked_by"
-    t.string   "queue"
-    t.datetime "created_at"
-    t.datetime "updated_at"
-    t.index ["priority", "run_at"], name: "delayed_jobs_priority", using: :btree
+  create_table "delayed_jobs", id: :serial, force: :cascade do |t|
+    t.integer "priority", default: 0, null: false
+    t.integer "attempts", default: 0, null: false
+    t.text "handler", null: false
+    t.text "last_error"
+    t.datetime "run_at", precision: nil
+    t.datetime "locked_at", precision: nil
+    t.datetime "failed_at", precision: nil
+    t.string "locked_by"
+    t.string "queue"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
+    t.index ["priority", "run_at"], name: "delayed_jobs_priority"
   end
 
-  create_table "kares", force: :cascade do |t|
-    t.string   "sku"
-    t.string   "title"
-    t.string   "full_title"
-    t.string   "desc"
-    t.string   "cat"
-    t.string   "charact"
-    t.decimal  "price"
-    t.integer  "quantity"
-    t.integer  "quantity_euro"
-    t.string   "specialty"
-    t.string   "image"
-    t.string   "url"
-    t.string   "brand"
-    t.datetime "created_at",    null: false
-    t.datetime "updated_at",    null: false
+  create_table "kares", id: :serial, force: :cascade do |t|
+    t.string "sku"
+    t.string "title"
+    t.string "full_title"
+    t.string "desc"
+    t.string "cat"
+    t.string "charact"
+    t.decimal "price"
+    t.integer "quantity"
+    t.integer "quantity_euro"
+    t.string "specialty"
+    t.string "image"
+    t.string "url"
+    t.string "brand"
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
+    t.string "status", default: "new", null: false
   end
 
-  create_table "products", force: :cascade do |t|
-    t.string   "sku"
-    t.string   "title"
-    t.string   "desc"
-    t.string   "cat"
-    t.string   "charact"
-    t.string   "charact_gab"
-    t.decimal  "oldprice"
-    t.decimal  "price"
-    t.integer  "quantity"
-    t.string   "image"
-    t.string   "url"
-    t.datetime "created_at",  null: false
-    t.datetime "updated_at",  null: false
+  create_table "products", id: :serial, force: :cascade do |t|
+    t.string "sku"
+    t.string "title"
+    t.string "desc"
+    t.string "cat"
+    t.string "charact"
+    t.string "charact_gab"
+    t.decimal "oldprice"
+    t.decimal "price"
+    t.integer "quantity"
+    t.string "image"
+    t.string "url"
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
+    t.string "status", default: "new", null: false
   end
 
-  create_table "users", force: :cascade do |t|
-    t.string   "email",                  default: "", null: false
-    t.string   "encrypted_password",     default: "", null: false
-    t.string   "reset_password_token"
-    t.datetime "reset_password_sent_at"
-    t.datetime "remember_created_at"
-    t.datetime "created_at",                          null: false
-    t.datetime "updated_at",                          null: false
-    t.string   "name"
-    t.index ["email"], name: "index_users_on_email", unique: true, using: :btree
-    t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true, using: :btree
+  create_table "users", id: :serial, force: :cascade do |t|
+    t.string "email", default: "", null: false
+    t.string "encrypted_password", default: "", null: false
+    t.string "reset_password_token"
+    t.datetime "reset_password_sent_at", precision: nil
+    t.datetime "remember_created_at", precision: nil
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
+    t.string "name"
+    t.string "role"
+    t.index ["email"], name: "index_users_on_email", unique: true
+    t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end
 
 end
